+++ test/ClusterE2E/Basic.hs
@@ -3,12 +3,14 @@
 module ClusterE2E.Basic (spec) where
 
 import           ClusterE2E.Utils
-import           Control.Exception             (bracket)
+import           Control.Exception             (IOException, bracket)
 import qualified Data.ByteString.Char8         as BS8
+import           Data.List                     (isInfixOf)
 import           Database.Redis.Cluster        (calculateSlot)
 import           Database.Redis.Cluster.Client (closeClusterClient)
 import           Database.Redis.Command        (RedisCommands (..))
 import           Database.Redis.Resp           (RespData (..))
+import           System.IO.Error               (ioeGetErrorString)
 import           Test.Hspec
 
 spec :: Spec
@@ -116,10 +118,8 @@ spec = describe "Basic cluster operations" $ do
     it "returns WRONGTYPE error for wrong command on key type" $ do
       bracket createTestClusterClient closeClusterClient $ \client -> do
         _ <- runCmd_ client $ set "cluster:wrongtype" "value"
-        result <- runCmd client $ lpush "cluster:wrongtype" ["item"]
-        case result of
-          RespError err -> BS8.isInfixOf "WRONGTYPE" err `shouldBe` True
-          other -> expectationFailure $ "Expected WRONGTYPE error, got: " ++ show other
+        runCmd_ client (lpush "cluster:wrongtype" ["item"])
+          `shouldThrow` isWrongTypeError
 
     it "DEL removes keys in cluster" $ do
       bracket createTestClusterClient closeClusterClient $ \client -> do
@@ -129,3 +129,6 @@ spec = describe "Basic cluster operations" $ do
         _ <- runCmd_ client $ del ["cluster:deltest"]
         afterDel <- runCmd client $ get "cluster:deltest"
         afterDel `shouldBe` RespNullBulkString
+
+isWrongTypeError :: IOException -> Bool
+isWrongTypeError = isInfixOf "WRONGTYPE" . ioeGetErrorString